api/requirements.txt
anyio==3.6.1
autopep8==1.7.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
firebase-admin==5.4.0
google-api-core==2.10.1
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpcio==1.49.1
grpcio-status==1.49.1
h11==0.14.0
httplib2==0.20.4
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.1.4
MarkupSafe==2.1.1
msgpack==1.0.4
orjson==3.8.0
proto-plus==1.22.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.10.2
PyMySQL==1.0.2
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
ujson==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
